requirements.txt
python-dotenv
playwright
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz